Backend/requirements.txt
async-timeout==5.0.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
future==1.0.0
gevent==24.10.2
greenlet==3.1.1
hiredis==3.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.2.3
orjson==3.10.18
pandas==2.2.3
pickle-mixin==1.0.2
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
pyarrow==19.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==6.2.0
requests==2.32.3
requests-file==2.1.0
scipy==1.15.2
six==1.16.0
textwrap3==0.9.2
tldextract==5.1.2
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.1.0
gunicorn==23.0.0
celery==5.3.6